# Conflicts:
#	modules/javafx.base/src/main/java/javafx/beans/property/ListPropertyBase.java
#	modules/javafx.base/src/main/java/javafx/beans/property/SetPropertyBase.java
#	modules/javafx.base/src/test/java/test/javafx/beans/property/SetPropertyBaseTest.java

+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/monocle/LinuxTouchTransform.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014, 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2014, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -148,7 +148,7 @@ class LinuxTouchTransform {
                     "monocle.input." + product + ".max" + axisName,
                     caps.getMaximum());
             translates[index] = -minimum;
-            scalars[index] = ((double) (range)) / (maximum - minimum);
+            scalars[index] = range / (maximum - minimum);
             return null;
         });
     }
